requirements.txt
pdfkit==0.6.1
click==7.1.1
Jinja2==2.11.1
PyYAML==5.3